+++ src/test/java/school/faang/user_service/service/event/EventServiceTest.java
@@ -1,10 +1,15 @@
 package school.faang.user_service.service.event;
 
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.junit.jupiter.MockitoSettings;
+import org.mockito.quality.Strictness;
 import school.faang.user_service.dto.event.EventDto;
 import school.faang.user_service.entity.Skill;
 import school.faang.user_service.entity.User;
@@ -14,13 +19,17 @@ import school.faang.user_service.mapper.event.EventMapper;
 import school.faang.user_service.repository.event.EventRepository;
 import school.faang.user_service.validator.event.EventValidator;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
+@MockitoSettings(strictness = Strictness.LENIENT)
 class EventServiceTest {
     @Mock
     private EventMapper eventMapper;
@@ -118,6 +127,72 @@ class EventServiceTest {
 
         assertThrows(DataValidationException.class,
                 () -> eventService.create(eventDto));
+    }
+
+    @Test
+    @DisplayName("Неуспешное получение созданных событий по Id юзера")
+    public void testFailedGetOwnedEventsById() {
+        User user1 = User.builder().id(1L).active(true).build();
+        long userId = user1.getId();
+        List<Event> emptyEvents = new ArrayList<>();
+        Mockito.when(eventRepository.findAllByUserId(userId)).thenReturn(Collections.emptyList());
+
+        Assertions.assertIterableEquals(emptyEvents, eventService.getOwnedEvents(userId));
+    }
+
+    @Test
+    @DisplayName("Успешное получение всех событий по верному Id пользователя")
+    public void testGetParticipatedEventsByUserId() {
+        long userId = 1L;
+
+        List<Event> mockEvents = List.of(
+                Event.builder()
+                        .id(21L)
+                        .title("EventOne")
+                        .maxAttendees(2)
+                        .build(),
+                Event.builder()
+                        .id(22L)
+                        .title("EventTwo")
+                        .maxAttendees(2)
+                        .build()
+        );
+        when(eventRepository.findParticipatedEventsByUserId(userId)).thenReturn(mockEvents);
+        List<Event> events = eventService.getParticipatedEventsByUserId(userId);
+
+        verify(eventRepository, times(1)).findParticipatedEventsByUserId(userId);
+        assertEquals(mockEvents, events);
+
+    }
+
+    @Test
+    @DisplayName("Неуспешный поиск события по неверному Id")
+    public void testFailedGetEventByIncorrectId() {
+        long wrongId = 11L;
+        Mockito.when(eventRepository.findById(wrongId)).thenReturn(Optional.empty());
+
+        assertThrows(DataValidationException.class, () -> eventService.getEvent(wrongId));
+    }
+
+    @Test
+    @DisplayName("Успешное удаление события по верному Id")
+    public void testSuccessDeleteEventById() {
+        Event eventDelete = Event.builder()
+                .id(5L)
+                .maxAttendees(5)
+                .build();
+        long eventId = eventDelete.getId();
+
+        eventService.deleteEvent(eventId);
+        Mockito.verify(eventRepository, times(1)).deleteById(eventId);
+    }
+
+    @Test
+    @DisplayName("Неуспешное удаление события по неверному Id")
+    public void testFailedDeleteEventByIncorrectId() {
+        long wrongId = 15L;
+
+        Mockito.verify(eventRepository, Mockito.never()).deleteById(wrongId);
     }
 
 }
